requirements.txt
pygments
pillow
httpx
fastapi
starlette[full]
typer[all]
uvicorn
flask